Generalize get_sample_type
